+++ routes/api/contacts.js
@@ -2,16 +2,19 @@ const express = require("express");
 
 const ctrl = require("../../controllers/contacts");
 
+const { validateBody } = require("../../middlewares");
+const schemas = require("../../schemas/contacts");
+
 const router = express.Router();
 
 router.get("/", ctrl.getAll);
 
 router.get("/:contactId", ctrl.getById);
 
-router.post("/", ctrl.add);
+router.post("/", validateBody(schemas.addSchema), ctrl.add);
 
 router.delete("/:contactId", ctrl.deleteById);
 
-router.put("/:contactId", ctrl.updateById);
+router.put("/:contactId", validateBody(schemas.addSchema), ctrl.updateById);
 
 module.exports = router;
